+++ spec/api/endpoints/teams_endpoint_spec.rb
@@ -48,11 +48,12 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
     context 'register a bot via oauth v2' do
       before do
         SlackRubyBotServer.config.oauth_version = :v2
-        oauth_access = Slack::Messages::Message.new({
-                                                      'access_token' => 'access_token',
-                                                      'authed_user' => { 'id' => 'user_id' },
-                                                      'team' => { 'id' => 'team_id', 'name' => 'team_name' }
-                                                    })
+        oauth_access = Slack::Messages::Message.new(
+          'access_token' => 'access_token',
+          'scope' => 'commands,incoming-webhook',
+          'authed_user' => { 'id' => 'user_id' },
+          'team' => { 'id' => 'team_id', 'name' => 'team_name' }
+        )
         ENV['SLACK_CLIENT_ID'] = 'client_id'
         ENV['SLACK_CLIENT_SECRET'] = 'client_secret'
         allow_any_instance_of(Slack::Web::Client).to receive(:oauth_v2_access).with(
@@ -75,6 +76,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.name).to eq 'team_name'
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v2'
+          expect(team.oauth_scope).to eq 'commands,incoming-webhook'
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
           expect(team.bot_user_id).to be nil
@@ -96,6 +99,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v2'
+          expect(team.oauth_scope).to eq 'commands,incoming-webhook'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
@@ -113,6 +118,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v2'
+          expect(team.oauth_scope).to eq 'commands,incoming-webhook'
           expect(team.active).to be true
           expect(team.bot_user_id).to be nil
           expect(team.activated_user_id).to eq 'user_id'
@@ -136,6 +143,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v2'
+          expect(team.oauth_scope).to eq 'commands,incoming-webhook'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
@@ -147,12 +156,13 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
     context 'register a bot via oauth v1' do
       before do
         SlackRubyBotServer.config.oauth_version = :v1
-        oauth_access = Slack::Messages::Message.new({
-                                                      'access_token' => 'access_token',
-                                                      'user_id' => 'user_id',
-                                                      'team_id' => 'team_id',
-                                                      'team_name' => 'team_name'
-                                                    })
+        oauth_access = Slack::Messages::Message.new(
+          'access_token' => 'access_token',
+          'scope' => 'incoming-webhook,commands,bot',
+          'user_id' => 'user_id',
+          'team_id' => 'team_id',
+          'team_name' => 'team_name'
+        )
         ENV['SLACK_CLIENT_ID'] = 'client_id'
         ENV['SLACK_CLIENT_SECRET'] = 'client_secret'
         allow_any_instance_of(Slack::Web::Client).to receive(:oauth_access).with(
@@ -175,6 +185,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.name).to eq 'team_name'
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
           expect(team.bot_user_id).to be nil
@@ -196,6 +208,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
@@ -213,6 +227,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.bot_user_id).to be nil
           expect(team.activated_user_id).to eq 'user_id'
@@ -236,6 +252,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'access_token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
@@ -247,16 +265,17 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
     context 'register a legacy bot' do
       before do
         SlackRubyBotServer.config.oauth_version = :v1
-        oauth_access = Slack::Messages::Message.new({
-                                                      'bot' => {
-                                                        'bot_access_token' => 'token',
-                                                        'bot_user_id' => 'bot_user_id'
-                                                      },
-                                                      'access_token' => 'access_token',
-                                                      'user_id' => 'user_id',
-                                                      'team_id' => 'team_id',
-                                                      'team_name' => 'team_name'
-                                                    })
+        oauth_access = Slack::Messages::Message.new(
+          'bot' => {
+            'bot_access_token' => 'token',
+            'bot_user_id' => 'bot_user_id'
+          },
+          'access_token' => 'access_token',
+          'scope' => 'incoming-webhook,commands,bot',
+          'user_id' => 'user_id',
+          'team_id' => 'team_id',
+          'team_name' => 'team_name'
+        )
         ENV['SLACK_CLIENT_ID'] = 'client_id'
         ENV['SLACK_CLIENT_SECRET'] = 'client_secret'
         allow_any_instance_of(Slack::Web::Client).to receive(:oauth_access).with(
@@ -279,6 +298,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.name).to eq 'team_name'
           team = Team.find(team.id)
           expect(team.token).to eq 'token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
           expect(team.bot_user_id).to eq 'bot_user_id'
@@ -300,6 +321,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'
@@ -317,6 +340,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.bot_user_id).to eq 'bot_user_id'
           expect(team.activated_user_id).to eq 'user_id'
@@ -340,6 +365,8 @@ describe SlackRubyBotServer::Api::Endpoints::TeamsEndpoint do
           expect(team.active).to be true
           team = Team.find(team.id)
           expect(team.token).to eq 'token'
+          expect(team.oauth_version).to eq 'v1'
+          expect(team.oauth_scope).to eq 'incoming-webhook,commands,bot'
           expect(team.active).to be true
           expect(team.activated_user_access_token).to eq 'access_token'
           expect(team.activated_user_id).to eq 'user_id'