+++ spec/database_adapters/activerecord/schema.rb
@@ -8,6 +8,8 @@ ActiveRecord::Schema.define do
     t.string :name
     t.string :domain
     t.string :token
+    t.string :oauth_scope
+    t.string :oauth_version, default: 'v1', null: false
     t.string :bot_user_id
     t.string :activated_user_id
     t.string :activated_user_access_token