+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 #### 1.2.0 (Next)
 
+* [#133](https://github.com/slack-ruby/slack-ruby-bot-server/pull/133): Added `Team#oauth_version` and `#scope` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 #### 1.1.0 (2020/11/17)
 
-* [#132](https://github.com/slack-ruby/slack-ruby-bot-server/pull/132): Add support for OAuth v2 - [@dblock](https://github.com/dblock).
+* [#132](https://github.com/slack-ruby/slack-ruby-bot-server/pull/132): Added support for OAuth v2 - [@dblock](https://github.com/dblock).
 
 #### 1.0.0 (2020/11/15)